Move API docs generation scripts to a single directory
