+++ docs/config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  /**
+   * Valid languages to server-side render in production
+   */
+  LANGUAGES: ['en', 'zh', 'pt'],
+};
