+++ docs/scripts/apiDocsBuilder/buildApiUtils.ts
@@ -2,9 +2,9 @@ import fs from 'fs';
 import path from 'path';
 import kebabCase from 'lodash/kebabCase';
 import { getHeaders, getTitle } from '@mui/markdown';
-import { findPagesMarkdownNew } from 'docs/src/modules/utils/find';
-import { getLineFeed } from 'docs/scripts/helpers';
-import { replaceComponentLinks } from 'docs/src/modules/utils/replaceUrl';
+import { replaceComponentLinks } from './utils/replaceUrl';
+import findPagesMarkdownNew from './utils/findPagesMarkdown';
+import { getLineFeed } from '../helpers';
 
 const systemComponents = fs
   .readdirSync(path.resolve('packages', 'mui-system', 'src'))
