+++ docs/scripts/apiDocsBuilder/utils/defaultPropsHandler.js
