+++ docs/scripts/apiDocsBuilder/utils/generatePropTypeDescription.ts
@@ -1,6 +1,6 @@
 import * as recast from 'recast';
 import { parse as docgenParse, PropTypeDescriptor } from 'react-docgen';
-import { escapeCell } from './helpers';
+import escapeCell from './escapeCell';
 
 function getDeprecatedInfo(type: PropTypeDescriptor) {
   const marker = /deprecatedPropType\((\r*\n)*\s*PropTypes\./g;
