+++ docs/scripts/apiDocsBuilder/utils/replaceUrl.ts
