+++ docs/scripts/tsconfig.json
@@ -1,11 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": [
-    "../types/react-docgen.d.ts",
-    "apiDocsBuilder/buildApi.ts",
-    "apiDocsBuilder/ApiBuilders/**/*",
-    "i18n.js"
-  ],
+  "include": ["../types/react-docgen.d.ts", "apiDocsBuilder/**/*", "i18n.js"],
   "compilerOptions": {
     "allowJs": true,
     "isolatedModules": true,
