+++ docs/src/modules/utils/find.js
@@ -1,80 +1,6 @@
 const fs = require('fs');
 const path = require('path');
 
-/**
- * Returns the markdowns of the documentation in a flat array.
- * @param {string} [directory]
- * @param {Array<{ filename: string, pathname: string }>} [pagesMarkdown]
- * @returns {Array<{ filename: string, pathname: string }>}
- */
-function findPagesMarkdownNew(
-  directory = path.resolve(__dirname, '../../../data'),
-  pagesMarkdown = [],
-) {
-  const items = fs.readdirSync(directory);
-
-  items.forEach((item) => {
-    const itemPath = path.resolve(directory, item);
-
-    if (fs.statSync(itemPath).isDirectory()) {
-      findPagesMarkdownNew(itemPath, pagesMarkdown);
-      return;
-    }
-
-    if (!/\.md$/.test(item) || /-(zh|pt)\.md/.test(item)) {
-      // neglect translation markdown
-      return;
-    }
-
-    let pathname = itemPath
-      .replace(new RegExp(`\\${path.sep}`, 'g'), '/')
-      .replace(/^.*\/data/, '')
-      .replace('.md', '');
-
-    // Remove the last pathname segment.
-    pathname = pathname.split('/').slice(0, 4).join('/');
-
-    pagesMarkdown.push({
-      // Relative location in the path (URL) system.
-      pathname,
-      // Relative location in the file system.
-      filename: itemPath,
-    });
-  });
-
-  return pagesMarkdown;
-}
-
-const componentRegex = /^(Unstable_)?([A-Z][a-z]+)+\.(js|tsx)/;
-
-/**
- * Returns the component source in a flat array.
- * @param {string} directory
- * @param {Array<{ filename: string }>} components
- */
-function findComponents(directory, components = []) {
-  const items = fs.readdirSync(directory);
-
-  items.forEach((item) => {
-    const itemPath = path.resolve(directory, item);
-
-    if (fs.statSync(itemPath).isDirectory()) {
-      findComponents(itemPath, components);
-      return;
-    }
-
-    if (!componentRegex.test(item)) {
-      return;
-    }
-
-    components.push({
-      filename: itemPath,
-    });
-  });
-
-  return components;
-}
-
 const pageRegex = /(\.js|\.tsx)$/;
 const blackList = ['/.eslintrc', '/_document', '/_app'];
 
@@ -157,6 +83,4 @@ function findPages(
 
 module.exports = {
   findPages,
-  findPagesMarkdownNew,
-  findComponents,
 };
