+++ docs/src/modules/utils/helpers.ts
@@ -106,8 +106,3 @@ export function pathnameToLanguage(pathname: string): {
     canonicalPathname,
   };
 }
-
-export function escapeCell(value: string): string {
-  // As the pipe is use for the table structure
-  return value.replace(/</g, '&lt;').replace(/`&lt;/g, '`<').replace(/\|/g, '\\|');
-}
